pyproject.toml
[project]
name = "django-jotter"
dynamic = ["version"]
authors = [{ name = "Vince Veselosky", email = "vince@veselosky.com" }]
description = "A note-taking app for Django"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "django >= 4.2, < 5.3",
    "pillow >= 10.0.1",
    "easy-thumbnails >= 2.8.4",
    "django-taggit >= 5.0.1",
    "django-tinymce >= 3.6, < 3.7",
]

[project.optional-dependencies]
dev = [
    "coverage >= 7.2.2",
    "django-debug-toolbar",
    "django_extensions",
    "django-environ",
    "django-rich >= 1.5.0",
    "docutils >= 0.20.1",
    "ipython >= 8.12",
    "pip-tools >= 6.12.3",
]

[project.urls]
"Homepage" = "https://github.com/veselosky/django-jotter"
"Bug Tracker" = "https://github.com/veselosky/django-jotter/issues"

[build-system]
# With setuptools-scm, package data files (e.g. templates) that are tracked by git will
# be automatically detected and included. Without setuptools-scm, you would need to
# specify a file pattern in MANIFEST.in to collect them.
requires = ["setuptools>=65.0", "setuptools-scm>=7.0.5", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = { attr = "jotter.__version__" }
